starter/ml/model.py
import pathlib
import pickle
import numpy as np
from numpy import typing as npt
from sklearn.metrics import fbeta_score, precision_score, recall_score
from sklearn.ensemble import RandomForestClassifier


# Optional: implement hyperparameter tuning.
def train_model(
    X_train: npt.NDArray[np.float32], y_train: npt.NDArray[np.float32]
) -> RandomForestClassifier:
    """
    Trains a machine learning model and returns it.

    Inputs
    ------
    X_train : np.array
        Training data.
    y_train : np.array
        Labels.
    Returns
    -------
    model
        Trained machine learning model.
    """

    clf = RandomForestClassifier()
    clf.fit(X_train, y_train)
    return clf


def compute_model_metrics(y, preds):
    """
    Validates the trained machine learning model using precision, recall, and F1.

    Inputs
    ------
    y : np.array
        Known labels, binarized.
    preds : np.array
        Predicted labels, binarized.
    Returns
    -------
    precision : float
    recall : float
    fbeta : float
    """
    fbeta = fbeta_score(y, preds, beta=1, zero_division=1)
    precision = precision_score(y, preds, zero_division=1)
    recall = recall_score(y, preds, zero_division=1)
    return precision, recall, fbeta


def inference(
    model: RandomForestClassifier, X: npt.NDArray[np.float32]
) -> npt.NDArray[np.float32]:
    """Run model inferences and return the predictions.

    Inputs
    ------
    model : sklearn.ensemble.RandomForestClassifier
        Trained machine learning model.
    X : np.array
        Data used for prediction.
    Returns
    -------
    preds : np.array
        Predictions from the model.
    """
    return model.predict(X)


def save(model: RandomForestClassifier, model_path: pathlib.Path):
    with open(model_path, "wb") as fp:
        pickle.dump(model, fp)


def load(model_path: pathlib.Path):
    with open(model_path, "rb") as fp:
        model = pickle.load(fp)

    return model
